- fixed: 'level.frozen' still needs to be writable and have an effect.
# Conflicts:
#	src/gamedata/g_mapinfo.h
#	src/scripting/vmthunks.cpp

# Conflicts:
#	src/g_levellocals.h

+++ src/g_levellocals.h
@@ -163,7 +163,7 @@ struct FLevelLocals : public FLevelData
 
 	bool		FromSnapshot;			// The current map was restored from a snapshot
 	bool		HasHeightSecs;			// true if some Transfer_Heights effects are present in the map. If this is false, some checks in the renderer can be shortcut.
-	uint8_t		frozenstate;
+	int			frozenstate;
 
 	double		teamdamage;
